Update evaluate.py
"The original file would cause the statistical_parity_difference parameter to be overwritten by the value of the f1 parameter, which may be related to the version of the related library used. This situation will occur when operating under Windows OS according to the Requirements_win.txt file. This adjustment to the evaluate.py file: 1. Fixes the issue where the statistical_parity_difference parameter is overwritten by the value of the f1 parameter under certain configurations. 2. The log transformation of the theil_index parameter has been commented out and is no longer used as the default output form (to be consistent with the format of the data in the referenced .csv file). You can choose freely."